+++ omego/artifacts.py
@@ -360,7 +360,7 @@ class ReleaseArtifacts(ArtifactsList):
         elif re.match('[0-9]+|latest$', args.branch):
             dl_url = self.follow_latest_redirect(args)
 
-        dl_icever = self.read_downloads(dl_url)
+        dl_icever = self.read_downloads(dl_url + 'artifacts/')
         if not args.ice:
             ice_ver = sorted(dl_icever.keys())[-1]
         else:
@@ -431,7 +431,8 @@ class ReleaseArtifacts(ArtifactsList):
 
 class DownloadCommand(Command):
     """
-    Download an OMERO artifact from a CI server.
+    Download an OMERO artifact from either a downloads or a Continuous
+    Integration server.
     """
 
     NAME = "download"